requirements.txt
transformers==2.11.0
torch==1.4.0
tensorboard==2.0